[refactor] 식당 별점 형 변환

+++ src/main/java/com/udoolleh/backend/entity/Restaurant.java
@@ -31,7 +31,7 @@ public class Restaurant {
     private String address;
 
     @Column(name = "total_grade")
-    private Float totalGrade;
+    private Double totalGrade;
 
     @OneToMany(mappedBy = "restaurant")
     private List<Menu> menuList = new ArrayList<>();
@@ -43,7 +43,7 @@ public class Restaurant {
     private List<Photo> photoList = new ArrayList<>();
 
     @Builder
-    public Restaurant(String name, String category,PlaceType placeType, String address, Float totalGrade){
+    public Restaurant(String name, String category,PlaceType placeType, String address, Double totalGrade){
         this.name = name;
         this.category = category;
         this.placeType = placeType;
